requirement.txt
django
djangorestframework
corsheaders
djangorestframework-simplejwt
PyJWT